requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
easy-thumbnails==2.8.5
gunicorn==21.2.0
idna==3.4
mailjet-rest==1.3.4
ngrok==0.12.0
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.6
PyJWT==2.8.0
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.3
